requirements.txt
python-telegram-bot==20.7
google-generativeai==0.7.2
supabase==2.5.0
python-dotenv==1.0.1
